+++ src/collective/documentgenerator/tests/test_setup.py
@@ -1,9 +1,11 @@
 # -*- coding: utf-8 -*-
 
 from collective.documentgenerator.content.pod_template import POD_TEMPLATE_TYPES
+from collective.documentgenerator.interfaces import ICollectiveDocumentGeneratorLayer
 from collective.documentgenerator.testing import NAKED_PLONE_INTEGRATION
 from collective.documentgenerator.testing import POD_TEMPLATE_INTEGRATION
-from plone import api
+from plone.browserlayer import utils
+from Products.CMFPlone.utils import get_installer
 
 import unittest
 
@@ -20,29 +22,48 @@ class TestInstallDependencies(unittest.TestCase):
         dexterity should be installed when we install documentgenerator
         """
         dependencies = self.portal.portal_setup.getProfileDependencyChain("collective.documentgenerator:default")
-        self.assertTrue(u"profile-plone.app.dexterity:default" in dependencies)
+        self.assertIn(u"profile-plone.app.dexterity:default", dependencies)
 
     def test_z3cformdatagridfield_is_dependency_of_documentgenerator(self):
         """
         z3cform.datagridfield should be installed when we install documentgenerator
         """
         dependencies = self.portal.portal_setup.getProfileDependencyChain("collective.documentgenerator:default")
-        self.assertTrue(u"profile-collective.z3cform.datagridfield:default" in dependencies)
+        self.assertIn(u"profile-collective.z3cform.datagridfield:default", dependencies)
 
 
 class TestSetup(unittest.TestCase):
 
     layer = POD_TEMPLATE_INTEGRATION
 
+    def setUp(self):
+        """Custom shared utility setup for tests."""
+        self.portal = self.layer["portal"]
+        self.installer = get_installer(self.portal)
+
     def test_pod_templates_folder_allowed_types(self):
 
-        portal = api.portal.get()
-        pod_folder = portal.podtemplates
+        pod_folder = self.portal.podtemplates
 
         allowed_types = [fti.__name__ for fti in pod_folder.allowedContentTypes()]
 
         msg = "pod folder allowed content types should only be the ones from documentgenerator"
-        self.assertTrue(len(allowed_types) == len(POD_TEMPLATE_TYPES), msg)
+        self.assertEqual(len(allowed_types), len(POD_TEMPLATE_TYPES), msg)
         for portal_type in POD_TEMPLATE_TYPES:
             msg = "type '{}' should be in pod folder allowed types".format(portal_type)
-            self.assertTrue(portal_type in allowed_types, msg)
+            self.assertIn(portal_type, allowed_types, msg)
+
+    def test_product_installed(self):
+        """Test if collective.documentgenerator is installed with portal_quickinstaller."""
+        self.assertTrue(self.installer.is_product_installed("collective.documentgenerator"))
+
+    def test_uninstall(self):
+        """Test if collective.documentgenerator is cleanly uninstalled."""
+        self.installer.uninstall_product("collective.documentgenerator")
+        self.assertFalse(self.installer.is_product_installed("collective.documentgenerator"))
+
+    def test_browserlayer(self):
+        """Test that ICollectiveDocumentGeneratorLayer is registered."""
+        self.assertIn(ICollectiveDocumentGeneratorLayer, utils.registered_layers())
+        self.installer.uninstall_product("collective.documentgenerator")
+        self.assertNotIn(ICollectiveDocumentGeneratorLayer, utils.registered_layers())